feat(new_bin_bench): Implement file parameter in benches. Revert __ before index
