+++ iai-callgrind-macros/src/common.rs
@@ -1,13 +1,28 @@
+use std::fs::File as StdFile;
+use std::io::{BufRead, BufReader};
+use std::path::PathBuf;
+
 use proc_macro2::{Span, TokenStream};
 use proc_macro_error::{abort, emit_error};
 use quote::{format_ident, quote, quote_spanned, ToTokens, TokenStreamExt};
 use syn::parse::Parse;
 use syn::spanned::Spanned;
-use syn::{parse2, Expr, ExprArray, ExprPath, Ident, MetaList, MetaNameValue, Token};
+use syn::{
+    parse2, parse_quote_spanned, Expr, ExprArray, ExprPath, Ident, LitStr, MetaList, MetaNameValue,
+    Token,
+};
+
+use crate::CargoMetadata;
 
 /// This struct reflects the `args` parameter of the `#[bench]` attribute
 #[derive(Debug, Default, Clone)]
 pub struct Args(Option<(Span, Vec<Expr>)>);
+
+#[derive(Debug, Clone)]
+pub struct Bench {
+    pub id: Ident,
+    pub args: Args,
+}
 
 /// The `config` parameter of the `#[bench]` or `#[benches]` attribute
 #[derive(Debug, Default, Clone)]
@@ -17,9 +32,15 @@ pub struct BenchConfig(pub Option<Expr>);
 #[derive(Debug, Clone, Default)]
 pub struct BenchesArgs(pub Option<Vec<Args>>);
 
+/// The `file` parameter of the `#[benches]` attribute
+#[derive(Debug, Default, Clone)]
+pub struct File(pub Option<LitStr>);
+
+/// The `setup` parameter
 #[derive(Debug, Default, Clone)]
 pub struct Setup(pub Option<ExprPath>);
 
+/// The `teardown` parameter
 #[derive(Debug, Default, Clone)]
 pub struct Teardown(pub Option<ExprPath>);
 
@@ -150,6 +171,72 @@ impl ToTokens for Args {
     }
 }
 
+impl Bench {
+    /// Return a vector of [`Bench`] parsing the [`File`] or [`BenchesArgs`] if present
+    ///
+    /// # Aborts
+    ///
+    /// If there are args in [`BenchesArgs`] and a [`File`] present. We can deal with only one them.
+    pub(crate) fn from_benches_attribute(
+        id: &Ident,
+        args: BenchesArgs,
+        file: &File,
+        cargo_meta: Option<&CargoMetadata>,
+        has_setup: bool,
+        expected_num_args: usize,
+    ) -> Vec<Self> {
+        match (&file.0, args.is_some()) {
+            (Some(literal), true) => {
+                abort!(
+                    literal.span(),
+                    "Only one parameter of `file` or `args` can be present"
+                );
+            }
+            (None, true) => args
+                .finalize()
+                .enumerate()
+                .map(|(index, args)| {
+                    args.check_num_arguments(expected_num_args, has_setup);
+                    let id = format_indexed_ident(id, index);
+                    Bench { id, args }
+                })
+                .collect(),
+            (Some(literal), false) => {
+                if !(expected_num_args == 1 || has_setup) {
+                    abort!(
+                        literal,
+                        "The benchmark function should take exactly one `String` argument if the file parameter is present";
+                        help = "fn benchmark_function(line: String) ..."
+                    )
+                }
+
+                let strings = file.read(cargo_meta);
+                if strings.is_empty() {
+                    abort!(literal, "The provided file '{}' was empty", literal.value());
+                }
+
+                let mut benches = vec![];
+                for (index, string) in strings.iter().enumerate() {
+                    let id = format_indexed_ident(id, index);
+                    let expr = if string.is_empty() {
+                        parse_quote_spanned! { literal.span() => String::new() }
+                    } else {
+                        parse_quote_spanned! { literal.span() => String::from(#string) }
+                    };
+                    let args = Args::new(literal.span(), vec![expr]);
+                    benches.push(Bench { id, args });
+                }
+                benches
+            }
+            // Cover the case when no arguments were present for example `#[benches::id()]`,
+            (None, false) => vec![Bench {
+                id: id.clone(),
+                args: Args::default(),
+            }],
+        }
+    }
+}
+
 impl BenchesArgs {
     pub fn is_some(&self) -> bool {
         self.0.is_some()
@@ -232,6 +319,67 @@ impl BenchConfig {
     }
 }
 
+impl File {
+    pub fn parse_pair(&mut self, pair: &MetaNameValue) {
+        if self.0.is_none() {
+            if let Expr::Lit(literal) = &pair.value {
+                self.0 = Some(parse2::<LitStr>(literal.to_token_stream()).unwrap());
+            } else {
+                abort!(
+                    pair.value, "Invalid value for `file`";
+                    help = "The `file` argument needs a literal string containing the path to an existing file at compile time";
+                    note = "`file = \"benches/some_fixture\"`"
+                );
+            }
+        } else {
+            emit_error!(
+                pair, "Duplicate argument: `file`";
+                help = "`file` is allowed only once"
+            );
+        }
+    }
+
+    /// Read this [`File`] and return all its lines
+    ///
+    /// # Panics
+    ///
+    /// Panics if there is no path present
+    pub(crate) fn read(&self, cargo_meta: Option<&CargoMetadata>) -> Vec<String> {
+        let expr = self.0.as_ref().expect("A file should be present");
+        let string = expr.value();
+        let mut path = PathBuf::from(&string);
+
+        if path.is_relative() {
+            if let Some(cargo_meta) = cargo_meta {
+                let root = PathBuf::from(&cargo_meta.workspace_root);
+                path = root.join(path);
+            }
+        }
+
+        let file = StdFile::open(&path)
+            .unwrap_or_else(|error| abort!(expr, "Error opening '{}': {}", path.display(), error));
+
+        let mut lines = vec![];
+        for (index, line) in BufReader::new(file).lines().enumerate() {
+            match line {
+                Ok(line) => {
+                    lines.push(line);
+                }
+                Err(error) => {
+                    abort!(
+                        self.0,
+                        "Error reading line {} in file '{}': {}",
+                        index + 1,
+                        path.display(),
+                        error
+                    );
+                }
+            }
+        }
+        lines
+    }
+}
+
 impl Setup {
     pub fn parse_pair(&mut self, pair: &MetaNameValue) {
         if self.0.is_none() {
@@ -306,10 +454,6 @@ impl Teardown {
     }
 }
 
-pub fn pretty_expr_path(expr: &ExprPath) -> String {
-    expr.to_token_stream().to_string().replace(' ', "")
-}
-
 pub fn format_ident(prefix: &str, ident: Option<&Ident>) -> Ident {
     if let Some(ident) = ident {
         format_ident!("{prefix}_{ident}")
@@ -319,7 +463,11 @@ pub fn format_ident(prefix: &str, ident: Option<&Ident>) -> Ident {
 }
 
 pub fn format_indexed_ident(ident: &Ident, index: usize) -> Ident {
-    format_ident!("{ident}__{index}")
+    format_ident!("{ident}_{index}")
+}
+
+pub fn pretty_expr_path(expr: &ExprPath) -> String {
+    expr.to_token_stream().to_string().replace(' ', "")
 }
 
 /// Truncate a utf-8 [`std::str`] to a given `len`
